# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  phone?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type EmergencyType = typeof EmergencyType[keyof typeof EmergencyType];


export const EmergencyType = {
  accident: 'accident',
  fire: 'fire',
  heart_attack: 'heart_attack',
  theft: 'theft',
} as const;

export type EmergencyAlertStatus = typeof EmergencyAlertStatus[keyof typeof EmergencyAlertStatus];


export const EmergencyAlertStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export interface EmergencyAlert {
  id: number;
  userId: number;
  type: EmergencyType;
  latitude?: number | null;
  longitude?: number | null;
  address?: string | null;
  status: EmergencyAlertStatus;
  createdAt: string;
}

export interface CreateEmergencyBody {
  type: EmergencyType;
  latitude?: number | null;
  longitude?: number | null;
  address?: string | null;
}

export type UpdateAlertStatusBodyStatus = typeof UpdateAlertStatusBodyStatus[keyof typeof UpdateAlertStatusBodyStatus];


export const UpdateAlertStatusBodyStatus = {
  active: 'active',
  resolved: 'resolved',
} as const;

export interface UpdateAlertStatusBody {
  status: UpdateAlertStatusBodyStatus;
}

export type NearbyLocationType = typeof NearbyLocationType[keyof typeof NearbyLocationType];


export const NearbyLocationType = {
  hospital: 'hospital',
  police: 'police',
  fire_station: 'fire_station',
} as const;

export interface NearbyLocation {
  id: string;
  name: string;
  type: NearbyLocationType;
  address: string;
  phone?: string | null;
  distance?: string | null;
  latitude: number;
  longitude: number;
}

export interface ListEmergenciesResponse {
  emergencies: EmergencyAlert[];
}

export interface ListNearbyResponse {
  locations: NearbyLocation[];
}

export interface MedicalProfile {
  id: number;
  userId: number;
  profileImage?: string | null;
  dob?: string | null;
  gender?: string | null;
  bloodGroup?: string | null;
  address?: string | null;
  allergies?: string | null;
  medicalConditions?: string | null;
  medications?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertProfileBody {
  profileImage?: string | null;
  dob?: string | null;
  gender?: string | null;
  bloodGroup?: string | null;
  address?: string | null;
  allergies?: string | null;
  medicalConditions?: string | null;
  medications?: string | null;
}

export interface EmergencyContact {
  id: number;
  userId: number;
  name: string;
  relationship: string;
  phone: string;
  priority: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  relationship: string;
  phone: string;
  priority: string;
}

export interface UpdateContactBody {
  name?: string;
  relationship?: string;
  phone?: string;
  priority?: string;
}

export interface ListContactsResponse {
  contacts: EmergencyContact[];
}

export interface ErrorResponse {
  error: string;
}

export type ListNearbyParams = {
latitude?: number;
longitude?: number;
};