# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Emergency Response API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: emergencies
    description: Emergency alert operations
  - name: nearby
    description: Nearby help locations
  - name: profile
    description: User medical profile
  - name: contacts
    description: Emergency contacts

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        phone:
          type: string
          nullable: true
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    EmergencyType:
      type: string
      enum:
        - accident
        - fire
        - heart_attack
        - theft

    EmergencyAlert:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          $ref: "#/components/schemas/EmergencyType"
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - active
            - resolved
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - status
        - createdAt

    CreateEmergencyBody:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/EmergencyType"
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
      required:
        - type

    UpdateAlertStatusBody:
      type: object
      properties:
        status:
          type: string
          enum:
            - active
            - resolved
      required:
        - status

    NearbyLocation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - hospital
            - police
            - fire_station
        address:
          type: string
        phone:
          type: string
          nullable: true
        distance:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
      required:
        - id
        - name
        - type
        - address
        - latitude
        - longitude

    ListEmergenciesResponse:
      type: object
      properties:
        emergencies:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyAlert"
      required:
        - emergencies

    ListNearbyResponse:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/NearbyLocation"
      required:
        - locations

    MedicalProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        profileImage:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        bloodGroup:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        allergies:
          type: string
          nullable: true
        medicalConditions:
          type: string
          nullable: true
        medications:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - createdAt
        - updatedAt

    UpsertProfileBody:
      type: object
      properties:
        profileImage:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        bloodGroup:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        allergies:
          type: string
          nullable: true
        medicalConditions:
          type: string
          nullable: true
        medications:
          type: string
          nullable: true

    EmergencyContact:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - relationship
        - phone
        - priority
        - createdAt

    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        priority:
          type: string
      required:
        - name
        - relationship
        - phone
        - priority

    UpdateContactBody:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        priority:
          type: string

    ListContactsResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyContact"
      required:
        - contacts

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emergencies:
    post:
      operationId: createEmergency
      tags: [emergencies]
      summary: Create emergency alert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmergencyBody"
      responses:
        "201":
          description: Emergency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listEmergencies
      tags: [emergencies]
      summary: List user emergency alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of emergencies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEmergenciesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emergencies/{id}/status:
    patch:
      operationId: updateEmergencyStatus
      tags: [emergencies]
      summary: Update emergency alert status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /nearby:
    get:
      operationId: listNearby
      tags: [nearby]
      summary: List nearby help locations
      parameters:
        - name: latitude
          in: query
          schema:
            type: number
        - name: longitude
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Nearby locations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListNearbyResponse"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user medical profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Medical profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update medical profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileBody"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List emergency contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListContactsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create emergency contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
        "400":
          description: Bad request or limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contacts/{id}:
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update emergency contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete emergency contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contact deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"